# contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Atomic Bank Operations API
  description: Production-safe banking microservice with atomic operations, auditable ledger, and event-driven architecture
  version: 1.0.0
  contact:
    name: Bank Operations Team
    email: ops@bank.example.com
  license:
    name: Proprietary
    url: https://bank.example.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.bank.example.com
    description: Production server
paths:
...
  /v1/accounts:
    post:
      summary: Create a new account
      operationId: createAccount
      tags:
        - Accounts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /v1/accounts/{accountId}:
    get:
      summary: Get account details
      operationId: getAccount
      tags:
        - Accounts
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
  /v1/accounts/{accountId}/withdraw:
    post:
      summary: Withdraw funds from account (atomic operation)
      operationId: withdraw
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
      responses:
        '200':
          description: Withdrawal processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid request or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency key conflict
  /v1/accounts/{accountId}/deposit:
    post:
      summary: Deposit funds into account (atomic operation)
      operationId: deposit
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Deposit processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
  /v1/accounts/{sourceAccountId}/transfer:
    post:
      summary: Transfer funds between accounts (atomic operation)
      operationId: transfer
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: sourceAccountId
          in: path
          required: true
          description: Source account UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid request or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
  /v1/accounts/{accountId}/transactions:
    get:
      summary: Get transaction history for account
      operationId: getTransactionHistory
      tags:
        - Transactions
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Account not found
...
components:
  schemas:
...
    CreateAccountRequest:
      type: object
      required:
        - holderName
        - initialBalance
        - currency
      properties:
        holderName:
          type: string
          description: Account holder name
          minLength: 1
        initialBalance:
          type: number
          format: decimal
          description: Initial account balance
          minimum: 0
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
          example: 'USD'
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Account ID
        holderName:
          type: string
          description: Account holder name
        balance:
          type: number
          format: decimal
          description: Current account balance
        currency:
          type: string
          description: Currency code
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
          description: Account status
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    WithdrawalRequest:
      type: object
      required:
        - amount
        - currency
        - idempotencyKey
      properties:
        amount:
          type: number
          format: decimal
          description: Withdrawal amount
          minimum: 0
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
        idempotencyKey:
          type: string
          format: uuid
          description: Unique idempotency key for this request
    DepositRequest:
      type: object
      required:
        - amount
        - currency
        - idempotencyKey
      properties:
        amount:
          type: number
          format: decimal
          description: Deposit amount
          minimum: 0
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
        idempotencyKey:
          type: string
          format: uuid
          description: Unique idempotency key for this request
    TransferRequest:
      type: object
      required:
        - destinationAccountId
        - amount
        - currency
        - idempotencyKey
      properties:
        destinationAccountId:
          type: string
          format: uuid
          description: Destination account ID
        amount:
          type: number
          format: decimal
          description: Transfer amount
          minimum: 0
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
        idempotencyKey:
          type: string
          format: uuid
          description: Unique idempotency key for this request
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Transaction ID
        type:
          type: string
          enum: [WITHDRAWAL, TRANSFER, DEPOSIT, FEE]
          description: Transaction type
        sourceAccountId:
          type: string
          format: uuid
          description: Source account ID (nullable for deposit)
        destinationAccountId:
          type: string
          format: uuid
          description: Destination account ID (nullable for withdrawal)
        status:
          type: string
          enum: [PENDING, POSTED, FAILED]
          description: Transaction status
        amount:
          type: number
          format: decimal
          description: Transaction amount
        currency:
          type: string
          description: Currency code
        idempotencyKey:
          type: string
          format: uuid
          description: Idempotency key
        initiatedBy:
          type: string
          description: User or service that initiated transaction
        createdAt:
          type: string
          format: date-time
          description: Transaction creation timestamp
        completedAt:
          type: string
          format: date-time
          description: Transaction completion timestamp (nullable if pending)
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        errorCode:
          type: string
          description: Application error code
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Request path
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /v1/auth/login
...